Respaldo

+++ frontend/JavaScript/Views/ViewBoards.mjs
@@ -1,6 +1,7 @@
 
-import { getByIdFunction, deleteFunction } from "../Model/Services/Board.Service.mjs";
+import { getByIdFunction, postFunction ,deleteFunction,putFunction } from "../Model/BoardModel/Board.Service.mjs";
 import { Url_Boards as url } from "../Utilities/config.mjs";
+import { eliminarLog } from "../Utilities/UtilsFunctions.mjs";
 
 export class Board {
 
@@ -24,6 +25,14 @@ export class Board {
 
         div.append(btnAdd)
 
+        btnAdd.addEventListener('click', async ()=>{
+
+            let name = prompt("Ingresa el nombre de la nueva tabla")
+
+            await postFunction(url,name)
+
+        })
+
         this.#data.data.forEach((item) => {
 
             // CreateElement
@@ -35,17 +44,19 @@ export class Board {
             const inputTextNameBoard = document.createElement("input")
             inputTextNameBoard.type ="text"
             inputTextNameBoard.value =item.name
+            inputTextNameBoard.readOnly = true
+            inputTextNameBoard.id = `btnName${item.id}`
 
             const verTablero = document.createElement("button")
             verTablero.className="verTablero"
             verTablero.id = item.id
 
             verTablero.addEventListener('click', ()=>{
-                alert(verTablero.id)
+                localStorage.setItem("id", verTablero.id)
             })
             
             const aEnlace = document.createElement("a")
-            aEnlace.href="./board.html" 
+            aEnlace.href="../../Html/board.html"
             aEnlace.innerHTML="Ver tablero"
 
             const buttonEdit = document.createElement("button")
@@ -53,9 +64,18 @@ export class Board {
             buttonEdit.className = "editarTablero"
             buttonEdit.innerHTML = "Editar"
 
-            buttonEdit.addEventListener('click', ()=>{
+            buttonEdit.addEventListener('click', async()=>{
+                
+                const id = verTablero.id
+                let btn = document.getElementById(`btnName${id}`)
+                btn.readOnly = false
+                btn.focus()
 
-                alert("El id a editar es: " + verTablero.id)
+                if(buttonEdit.innerHTML === "Cambiar"){
+                    await putFunction(url,id,inputTextNameBoard.value)
+                }
+
+                buttonEdit.innerHTML = "Cambiar"
 
             })
 
@@ -67,12 +87,27 @@ export class Board {
             buttonDel.addEventListener('click', async ()=>{
 
                 let data = await getByIdFunction(url,verTablero.id)
+                let taskData = data.data.task
 
-                console.log(data)
+                if (confirm(`seguro de eliminar este tablero? ${data.data.name}`)) {
 
-                alert("El id a eliminar es: " + verTablero.id)
+                    taskData.forEach( task =>{
 
-                await deleteFunction(url, verTablero.id)
+                        let logsData = task.logs
+    
+                        if(task.logs.length){
+                            logsData.forEach( log =>{ 
+    
+                                eliminarLog(log.id)
+        
+                            })
+                        }
+    
+                    })
+
+                    await deleteFunction(url, verTablero.id)
+                    
+                }               
 
             })
